CRM-17156 log user detail in queries to track down long queries
This is a possible approach to greater traceability on how long running queries are
generated. By logging the contact id the person can be advised (currently manually)
that their query has been killed & interrogated as to how they caused it

CRM-17156 log user id against queries fix

The 4.2 method for getting the user id for queries turned out to be causing a loop as it was being called too early in the instantiation process

CRM-17156 log user id against queries

This causes the logged in user to be recorded in mysql for figuring out the source of bad queries. It should be skipped during tests as it gets called early on in dev mode & causes a loop